+++ packages/dsbuntis/lib/src/plan.dart
@@ -1,52 +1,44 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:dsbuntis/src/day.dart';
-import 'package:dsbuntis/src/sub.dart';
+import 'package:untis/untis.dart' as untis;
 
-class Plan {
-  Day? day;
-  List<Substitution> subs;
-  String date;
+// TODO: add A LOT more documentation
+// TODO: this whole structure has to be rethought
+class Plan extends untis.Plan {
   String url;
   String previewUrl;
   Uint8List? preview;
-  // TODO: include more info from dsb
+  // TODO: more data from dsb
 
-  Plan(this.day, this.subs, this.date, this.url, this.previewUrl, this.preview);
+  Plan(untis.Day? day, List<untis.Substitution> subs, String date, this.url,
+      this.previewUrl, this.preview)
+      : super(day, subs, date);
+
+  Plan.from(untis.Plan p, this.url, this.previewUrl, this.preview)
+      : super(p.day, p.subs, p.date);
 
   Plan.fromJson(Map<String, dynamic> json)
-      : day = dayFromInt(json['day']),
-        date = json['date'],
-        subs = _subsFromJson(json['subs']),
-        url = json['url'],
+      : url = json['url'],
         previewUrl = json['preview_url'],
         preview = json.containsKey('preview')
             ? Uint8List.fromList(List<int>.from(json['preview']))
-            : null;
+            : null,
+        super.fromJson(json);
 
-  Map<String, dynamic> toJson() {
-    final m = <String, dynamic>{
-      'day': dayToInt(day),
-      'date': date,
-      'subs': _subsToJson(),
+  @override
+  Map<String, dynamic> toJson() => super.toJson()
+    ..addAll({
       'url': url,
       'preview_url': previewUrl,
-    };
-    if (preview != null) m['preview'] = preview;
-    return m;
-  }
-
-  dynamic _subsToJson() => subs.map((sub) => sub.toJson()).toList();
-
-  static List<Substitution> _subsFromJson(dynamic json) =>
-      json.map<Substitution>((s) => Substitution.fromJson(s)).toList();
+      if (preview != null) 'preview': preview,
+    });
 
   @override
   String toString([bool inclUrls = true]) => '$day${_cum(inclUrls)}: $subs';
   String _cum(bool iu) => iu ? '($url, $previewUrl)' : '';
 
-  static List plansToJson(List<Plan> plans) =>
+  static List plansToJson(Iterable<Plan> plans) =>
       plans.map((e) => e.toJson()).toList();
 
   static List<Plan> plansFromJson(dynamic plans) =>
@@ -59,8 +51,8 @@ class Plan {
       plansFromJson(jsonDecode(plans));
 
   static List<Plan> searchInPlans(
-    List<Plan> plans,
-    bool Function(Substitution) predicate,
+    Iterable<Plan> plans,
+    bool Function(untis.Substitution) predicate,
   ) =>
       plans
           .map((p) => Plan(p.day, p.subs.where(predicate).toList(), p.date,