+++ packages/untis/test/sort_test.dart
@@ -1,7 +1,7 @@
-import 'package:dsbuntis/dsbuntis.dart';
+import 'package:untis/untis.dart';
 import 'package:test/test.dart';
 
-import 'testlib.dart';
+import '../../../testlib.dart';
 
 TestCase sortTestCase(List<List> input, List<int> expected) => () async {
       if (input.length != expected.length) throw 'Invalid test.';